+++ src/main/java/org/auscope/portal/core/services/responses/csw/CSWRecordTransformerFactory.java
@@ -2,6 +2,8 @@ package org.auscope.portal.core.services.responses.csw;
 
 import javax.xml.parsers.ParserConfigurationException;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.auscope.portal.core.server.OgcServiceProviderType;
 import org.auscope.portal.core.services.PortalServiceException;
 import org.w3c.dom.Node;
@@ -13,15 +15,19 @@ import org.w3c.dom.Node;
  *
  */
 public class CSWRecordTransformerFactory {
+	
+	protected final Log logger = LogFactory.getLog(getClass());
+	
     /**
      * Creates a new instance of CSWRecordTransformer which will draw from the specified gmd:MD_Metadata Node representation as a template
      * 
      * @param mdMetadataNode
      * @return
      */
-    public CSWRecordTransformer newCSWRecordTransformer(Node mdMetadataNode) {
+    /*public CSWRecordTransformer newCSWRecordTransformer(Node mdMetadataNode) {
     	return new CSWRecordTransformer(mdMetadataNode);
     }
+    */
     
     /**
      * Creates a new instance of CSWRecordTransformer with CSW server type
@@ -31,14 +37,7 @@ public class CSWRecordTransformerFactory {
      * @return
      */
     public CSWRecordTransformer newCSWRecordTransformer(Node mdMetadataNode, OgcServiceProviderType serverType) {
-    	switch (serverType) {
-    	case PyCSW:
-    		return new PyCSWRecordTransformer(mdMetadataNode);
-    	case GeoServer:
-    		return new GeoserverCSWRecordTransformer(mdMetadataNode);
-    	default:
-    		return new CSWRecordTransformer(mdMetadataNode);
-    	}
+    	return new CSWRecordTransformer(mdMetadataNode, serverType);
     }
 
     /**
